LayoutTests/http/tests/cache/disk-cache/disk-cache-media-expected.txt

Tests that media resources loaded via XHR or Fetch is not cached.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


running 12 tests

--------Testing loads from disk cache--------
response headers: {"Cache-control":"max-age=0","Content-Type":"text/plain"}
response source: Network

response headers: {"Cache-control":"max-age=100","Content-Type":"text/plain"}
response source: Disk cache

response headers: {"Cache-control":"max-age=0","Content-Type":"video/mp4"}
response source: Network

response headers: {"Cache-control":"max-age=100","Content-Type":"video/mp4"}
response source: Network

response headers: {"Cache-control":"max-age=0","Content-Type":"audio/mp4"}
response source: Network

response headers: {"Cache-control":"max-age=100","Content-Type":"audio/mp4"}
response source: Network

response headers: {"Cache-control":"max-age=0","Content-Type":"text/plain"}
response source: Network (fetch)

response headers: {"Cache-control":"max-age=100","Content-Type":"text/plain"}
response source: Disk cache (fetch)

response headers: {"Cache-control":"max-age=0","Content-Type":"video/mp4"}
response source: Network (fetch)

response headers: {"Cache-control":"max-age=100","Content-Type":"video/mp4"}
response source: Network (fetch)

response headers: {"Cache-control":"max-age=0","Content-Type":"audio/mp4"}
response source: Network (fetch)

response headers: {"Cache-control":"max-age=100","Content-Type":"audio/mp4"}
response source: Network (fetch)

--------Testing loads through memory cache (XHR behavior)--------
response headers: {"Cache-control":"max-age=0","Content-Type":"text/plain"}
response source: Network

response headers: {"Cache-control":"max-age=100","Content-Type":"text/plain"}
response source: Memory cache

response headers: {"Cache-control":"max-age=0","Content-Type":"video/mp4"}
response source: Network

response headers: {"Cache-control":"max-age=100","Content-Type":"video/mp4"}
response source: Memory cache

response headers: {"Cache-control":"max-age=0","Content-Type":"audio/mp4"}
response source: Network

response headers: {"Cache-control":"max-age=100","Content-Type":"audio/mp4"}
response source: Memory cache

response headers: {"Cache-control":"max-age=0","Content-Type":"text/plain"}
response source: Network (fetch)

response headers: {"Cache-control":"max-age=100","Content-Type":"text/plain"}
response source: Disk cache (fetch)

response headers: {"Cache-control":"max-age=0","Content-Type":"video/mp4"}
response source: Network (fetch)

response headers: {"Cache-control":"max-age=100","Content-Type":"video/mp4"}
response source: Network (fetch)

response headers: {"Cache-control":"max-age=0","Content-Type":"audio/mp4"}
response source: Network (fetch)

response headers: {"Cache-control":"max-age=100","Content-Type":"audio/mp4"}
response source: Network (fetch)

--------Testing loads through memory cache (subresource behavior)--------
response headers: {"Cache-control":"max-age=0","Content-Type":"text/plain"}
response source: Network

response headers: {"Cache-control":"max-age=100","Content-Type":"text/plain"}
response source: Memory cache

response headers: {"Cache-control":"max-age=0","Content-Type":"video/mp4"}
response source: Network

response headers: {"Cache-control":"max-age=100","Content-Type":"video/mp4"}
response source: Memory cache

response headers: {"Cache-control":"max-age=0","Content-Type":"audio/mp4"}
response source: Network

response headers: {"Cache-control":"max-age=100","Content-Type":"audio/mp4"}
response source: Memory cache

response headers: {"Cache-control":"max-age=0","Content-Type":"text/plain"}
response source: Network (fetch)

response headers: {"Cache-control":"max-age=100","Content-Type":"text/plain"}
response source: Memory cache (fetch)

response headers: {"Cache-control":"max-age=0","Content-Type":"video/mp4"}
response source: Network (fetch)

response headers: {"Cache-control":"max-age=100","Content-Type":"video/mp4"}
response source: Memory cache (fetch)

response headers: {"Cache-control":"max-age=0","Content-Type":"audio/mp4"}
response source: Network (fetch)

response headers: {"Cache-control":"max-age=100","Content-Type":"audio/mp4"}
response source: Memory cache (fetch)

PASS successfullyParsed is true

TEST COMPLETE
